+++ src/libs/axios/instanse.ts
@@ -1,5 +1,5 @@
 import axios from 'axios';
-import { getSessionAccessToken } from '../../utils/auth-storage-manager';
+import { getSessionAccessToken } from '../../utils/token/auth-storage-manager';
 
 export const axiosInstanse = axios.create({
   baseURL: 'https://topdragon.co.kr',
